Fix printing of current examples
